Feat: restructure item-buffer to allow easy backup and restore

+++ src/buffer/item_buffer.cpp
@@ -17,7 +17,66 @@ struct EmptyPlaceHolder
     uint64_t bytePositionOfNextEmptyBlock = ITEM_BUFFER_UNDEFINE_POS;
 } __attribute__((packed));
 
+/**
+ * @brief constructor
+ */
 ItemBuffer::ItemBuffer() {}
+
+/**
+ * @brief destructor
+ */
+ItemBuffer::~ItemBuffer() {}
+
+/**
+ * @brief init basically a simple buffer without items.
+ *
+ * @param staticSize number of bytes to allocate
+ *
+ * @return true, if successful, else false
+ */
+bool
+ItemBuffer::initBuffer(const uint64_t staticSize)
+{
+    // allocate memory
+    const bool ret = initDataBlocks(0, 0, staticSize);
+    if(ret == false) {
+        return false;
+    }
+
+    return true;
+}
+
+/**
+ * @brief initialize a new item-buffer based on the payload of an old one
+ *
+ * @param data pointer to the data to import
+ * @param dataSize number of bytes of data
+ *
+ * @return true, if successful, else false if input is invalid
+ */
+bool
+ItemBuffer::initBuffer(const void* data, const uint64_t dataSize)
+{
+    // precheck
+    if(dataSize == 0
+            || data == nullptr)
+    {
+        return false;
+    }
+
+    // allocate blocks in buffer and fill with old data
+    Kitsunemimi::allocateBlocks_DataBuffer(buffer, calcBytesToBlocks(dataSize));
+    buffer.usedBufferSize = dataSize;
+    memcpy(buffer.data, data, dataSize);
+
+    // init pointer
+    metaData = static_cast<MetaData*>(buffer.data);
+    uint8_t* u8Data = static_cast<uint8_t*>(buffer.data);
+    staticData = &u8Data[sizeof(MetaData)];
+    itemData = &u8Data[sizeof(MetaData) + metaData->staticSize];
+
+    return true;
+}
 
 /**
  * @brief delete all items for the buffer
@@ -25,8 +84,12 @@ ItemBuffer::ItemBuffer() {}
 void
 ItemBuffer::deleteAll()
 {
+    if(metaData == nullptr) {
+        return;
+    }
+
     while(m_lock.test_and_set(std::memory_order_acquire)) { asm(""); }
-    for(uint64_t i = 0; i < itemCapacity; i++) {
+    for(uint64_t i = 0; i < metaData->itemCapacity; i++) {
         deleteItem(i);
     }
     m_lock.clear(std::memory_order_release);
@@ -46,22 +109,31 @@ ItemBuffer::initDataBlocks(const uint64_t numberOfItems,
                            const uint64_t staticSize)
 {
     // precheck
-    if(itemSize == 0) {
+    if(itemSize == 0
+            && staticSize == 0)
+    {
         return false;
     }
 
-    // update buffer-values
-    this->itemSize = itemSize;
-    this->itemCapacity = numberOfItems;
-    const uint64_t requiredBytes = (numberOfItems * itemSize) + staticSize;
-    const uint64_t requiredNumberOfBlocks = (requiredBytes / buffer.blockSize) + 1;
+    const uint64_t itemBytes = numberOfItems * itemSize;
+    const uint64_t requiredBytes = itemBytes + staticSize + sizeof(MetaData);
+    const uint64_t requiredNumberOfBlocks = calcBytesToBlocks(requiredBytes);
 
     // allocate blocks in buffer
     Kitsunemimi::allocateBlocks_DataBuffer(buffer, requiredNumberOfBlocks);
     buffer.usedBufferSize = requiredBytes;
 
-    staticData = buffer.data;
-    itemData = static_cast<uint8_t*>(buffer.data) + staticSize;
+    // init metadata object
+    metaData = static_cast<MetaData*>(buffer.data);
+    metaData[0] = MetaData();
+    metaData->itemSize = itemSize;
+    metaData->itemCapacity = numberOfItems;
+    metaData->staticSize = staticSize;
+
+    // init pointer
+    uint8_t* u8Data = static_cast<uint8_t*>(buffer.data);
+    staticData = &u8Data[sizeof(MetaData)];
+    itemData = &u8Data[sizeof(MetaData) + staticSize];
 
     return true;
 }
@@ -77,8 +149,10 @@ bool
 ItemBuffer::deleteItem(const uint64_t itemPos)
 {
     // precheck
-    if(itemPos >= itemCapacity
-            || numberOfItems == 0)
+    if(metaData == nullptr
+            || metaData->itemSize == 0
+            || itemPos >= metaData->itemCapacity
+            || metaData->numberOfItems == 0)
     {
         return false;
     }
@@ -87,7 +161,7 @@ ItemBuffer::deleteItem(const uint64_t itemPos)
     uint8_t* blockBegin = static_cast<uint8_t*>(itemData);
 
     // data of the position
-    const uint64_t currentBytePos = itemPos * itemSize;
+    const uint64_t currentBytePos = itemPos * metaData->itemSize;
     void* voidBuffer = static_cast<void*>(&blockBegin[currentBytePos]);
     EmptyPlaceHolder* placeHolder = static_cast<EmptyPlaceHolder*>(voidBuffer);
 
@@ -101,7 +175,7 @@ ItemBuffer::deleteItem(const uint64_t itemPos)
     placeHolder->status = ItemBuffer::DELETED_SECTION;
 
     // modify last place-holder
-    const uint64_t blockPosition = m_bytePositionOfLastEmptyBlock;
+    const uint64_t blockPosition = metaData->bytePositionOfLastEmptyBlock;
     if(blockPosition != ITEM_BUFFER_UNDEFINE_POS)
     {
         voidBuffer = static_cast<void*>(&blockBegin[blockPosition]);
@@ -110,12 +184,12 @@ ItemBuffer::deleteItem(const uint64_t itemPos)
     }
 
     // set global values
-    m_bytePositionOfLastEmptyBlock = currentBytePos;
-    if(m_bytePositionOfFirstEmptyBlock == ITEM_BUFFER_UNDEFINE_POS) {
-        m_bytePositionOfFirstEmptyBlock = currentBytePos;
+    metaData->bytePositionOfLastEmptyBlock = currentBytePos;
+    if(metaData->bytePositionOfFirstEmptyBlock == ITEM_BUFFER_UNDEFINE_POS) {
+        metaData->bytePositionOfFirstEmptyBlock = currentBytePos;
     }
 
-    numberOfItems--;
+    metaData->numberOfItems--;
 
     return true;
 }
@@ -129,7 +203,7 @@ uint64_t
 ItemBuffer::reuseItemPosition()
 {
     // get byte-position of free space, if exist
-    const uint64_t selectedPosition = m_bytePositionOfFirstEmptyBlock;
+    const uint64_t selectedPosition = metaData->bytePositionOfFirstEmptyBlock;
     if(selectedPosition == ITEM_BUFFER_UNDEFINE_POS) {
         return ITEM_BUFFER_UNDEFINE_POS;
     }
@@ -138,18 +212,18 @@ ItemBuffer::reuseItemPosition()
     uint8_t* blockBegin = static_cast<uint8_t*>(itemData);
     void* voidBuffer = static_cast<void*>(&blockBegin[selectedPosition]);
     EmptyPlaceHolder* secetedPlaceHolder = static_cast<EmptyPlaceHolder*>(voidBuffer);
-    m_bytePositionOfFirstEmptyBlock = secetedPlaceHolder->bytePositionOfNextEmptyBlock;
+    metaData->bytePositionOfFirstEmptyBlock = secetedPlaceHolder->bytePositionOfNextEmptyBlock;
 
     // reset pointer, if no more free spaces exist
-    if(m_bytePositionOfFirstEmptyBlock == ITEM_BUFFER_UNDEFINE_POS) {
-        m_bytePositionOfLastEmptyBlock = ITEM_BUFFER_UNDEFINE_POS;
+    if(metaData->bytePositionOfFirstEmptyBlock == ITEM_BUFFER_UNDEFINE_POS) {
+        metaData->bytePositionOfLastEmptyBlock = ITEM_BUFFER_UNDEFINE_POS;
     }
 
     // convert byte-position to item-position and return this
-    numberOfItems++;
-    assert(selectedPosition % itemSize == 0);
+    metaData->numberOfItems++;
+    assert(selectedPosition % metaData->itemSize == 0);
 
-    return selectedPosition / itemSize;
+    return selectedPosition / metaData->itemSize;
 }
 
 /**
@@ -169,16 +243,16 @@ ItemBuffer::reserveDynamicItem()
     // calculate size information
     const uint32_t blockSize = buffer.blockSize;
     const uint64_t numberOfBlocks = buffer.numberOfBlocks;
-    const uint64_t newNumberOfBlocks = (((itemCapacity + 1) * itemSize) / blockSize) + 1;
+    const uint64_t newNumberOfBlocks = (((metaData->itemCapacity + 1) * metaData->itemSize) / blockSize) + 1;
 
     // allocate a new block, if necessary
     if(numberOfBlocks < newNumberOfBlocks) {
         Kitsunemimi::allocateBlocks_DataBuffer(buffer, newNumberOfBlocks - numberOfBlocks);
     }
 
-    itemCapacity++;
+    metaData->itemCapacity++;
 
-    return itemCapacity-1;
+    return metaData->itemCapacity-1;
 }
 
 }
